chore: simplify
